+++ cmd/harbor/root/login_test.go
@@ -11,12 +11,13 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-package e2e
+package root_test
 
 import (
 	"testing"
 
 	"github.com/goharbor/harbor-cli/cmd/harbor/root"
+	helpers "github.com/goharbor/harbor-cli/test/helper"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -52,8 +53,8 @@ import (
 
 func Test_Login_Failure_WrongServer(t *testing.T) {
 	tempDir := t.TempDir()
-	data := Initialize(t, tempDir)
-	defer ConfigCleanup(t, data)
+	data := helpers.Initialize(t, tempDir)
+	defer helpers.ConfigCleanup(t, data)
 
 	cmd := root.LoginCommand()
 	cmd.SetArgs([]string{"wrongserver"})
@@ -68,8 +69,8 @@ func Test_Login_Failure_WrongServer(t *testing.T) {
 
 func Test_Login_Failure_WrongUsername(t *testing.T) {
 	tempDir := t.TempDir()
-	data := Initialize(t, tempDir)
-	defer ConfigCleanup(t, data)
+	data := helpers.Initialize(t, tempDir)
+	defer helpers.ConfigCleanup(t, data)
 
 	cmd := root.LoginCommand()
 	cmd.SetArgs([]string{"http://demo.goharbor.io"})
@@ -84,8 +85,8 @@ func Test_Login_Failure_WrongUsername(t *testing.T) {
 
 func Test_Login_Failure_WrongPassword(t *testing.T) {
 	tempDir := t.TempDir()
-	data := Initialize(t, tempDir)
-	defer ConfigCleanup(t, data)
+	data := helpers.Initialize(t, tempDir)
+	defer helpers.ConfigCleanup(t, data)
 
 	cmd := root.LoginCommand()
 	cmd.SetArgs([]string{"http://demo.goharbor.io"})
