+++ pkg/utils/config_test.go
@@ -11,88 +11,25 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-package e2e
+package utils_test
 
 import (
-	"os"
 	"path/filepath"
-	"sync"
 	"testing"
 
 	"github.com/goharbor/harbor-cli/cmd/harbor/root"
 	"github.com/goharbor/harbor-cli/pkg/utils"
+
+	helpers "github.com/goharbor/harbor-cli/test/helper"
 	"github.com/stretchr/testify/assert"
 )
 
-var envMutex sync.Mutex
-
-func safeSetEnv(key, value string) {
-	envMutex.Lock()
-	defer envMutex.Unlock()
-	os.Setenv(key, value)
-}
-
-func safeUnsetEnv(key string) {
-	envMutex.Lock()
-	defer envMutex.Unlock()
-	os.Unsetenv(key)
-}
-
-func ConfigCleanup(t *testing.T, data *utils.HarborData) {
-	if data != nil && data.ConfigPath != "" {
-		err := os.Remove(data.ConfigPath)
-		if err != nil && !os.IsNotExist(err) {
-			t.Fatalf("Failed to clean up test config file: %v", err)
-		}
-	}
-	if os.Getenv("XDG_CONFIG_HOME") != "" {
-		err := os.RemoveAll(os.Getenv("XDG_CONFIG_HOME"))
-		if err != nil {
-			t.Fatalf("Failed to clean up test config directory: %v", err)
-		}
-	}
-	if os.Getenv("XDG_DATA_HOME") != "" {
-		err := os.RemoveAll(os.Getenv("XDG_DATA_HOME"))
-		if err != nil {
-			t.Fatalf("Failed to clean up test data directory: %v", err)
-		}
-	}
-	safeUnsetEnv("HARBOR_CLI_CONFIG")
-	safeUnsetEnv("XDG_CONFIG_HOME")
-	safeUnsetEnv("XDG_DATA_HOME")
-	data = nil
-}
-
-func SetMockKeyring(t *testing.T) {
-	mockKeyring := utils.NewMockKeyring()
-	utils.SetKeyringProvider(mockKeyring)
-
-	t.Cleanup(func() {
-		utils.SetKeyringProvider(&utils.SystemKeyring{})
-	})
-}
-
-func Initialize(t *testing.T, tempDir string) *utils.HarborData {
-	utils.ConfigInitialization.Reset() // Reset sync.Once for the test
-	SetMockKeyring(t)
-	safeSetEnv("XDG_DATA_HOME", filepath.Join(tempDir, ".data"))
-	utils.InitConfig(filepath.Join(tempDir, ".config", "config.yaml"), true)
-	cds := root.RootCmd()
-	err := cds.Execute()
-	assert.NoError(t, err, "Expected no error for Root command")
-	assert.NoError(t, err, "Expected no error for Root command execution")
-
-	currentData, err := utils.GetCurrentHarborData()
-	assert.NoError(t, err, "Expected no error when fetching HarborData")
-	return currentData
-}
-
 func Test_Config_EnvVar(t *testing.T) {
 	utils.ConfigInitialization.Reset() // Reset sync.Once for the test
-	SetMockKeyring(t)
+	helpers.SetMockKeyring(t)
 	tempDir := t.TempDir()
-	safeSetEnv("HARBOR_CLI_CONFIG", filepath.Join(tempDir, "config.yaml"))
-	safeSetEnv("XDG_DATA_HOME", filepath.Join(tempDir, ".data"))
+	helpers.SafeSetEnv("HARBOR_CLI_CONFIG", filepath.Join(tempDir, "config.yaml"))
+	helpers.SafeSetEnv("XDG_DATA_HOME", filepath.Join(tempDir, ".data"))
 	utils.InitConfig("", false)
 	cds := root.RootCmd()
 	err := cds.Execute()
@@ -101,7 +38,7 @@ func Test_Config_EnvVar(t *testing.T) {
 
 	currentData, err := utils.GetCurrentHarborData()
 	assert.NoError(t, err, "Expected no error when fetching HarborData")
-	defer ConfigCleanup(t, currentData)
+	defer helpers.ConfigCleanup(t, currentData)
 
 	currentConfig, err := utils.GetCurrentHarborConfig()
 	assert.NoError(t, err, "Expected no error when fetching HarborConfig")
@@ -113,7 +50,7 @@ func Test_Config_EnvVar(t *testing.T) {
 
 func Test_Config_Vanilla(t *testing.T) {
 	utils.ConfigInitialization.Reset() // Reset sync.Once for the test
-	SetMockKeyring(t)
+	helpers.SetMockKeyring(t)
 	utils.InitConfig("", false)
 	cds := root.RootCmd()
 	err := cds.Execute()
@@ -121,7 +58,7 @@ func Test_Config_Vanilla(t *testing.T) {
 	assert.NoError(t, err, "Expected no error for Root command execution")
 	currentData, err := utils.GetCurrentHarborData()
 	assert.NoError(t, err, "Expected no error when fetching HarborData")
-	defer ConfigCleanup(t, currentData)
+	defer helpers.ConfigCleanup(t, currentData)
 
 	currentConfig, err := utils.GetCurrentHarborConfig()
 	assert.NoError(t, err, "Expected no error when fetching HarborConfig")
@@ -133,11 +70,11 @@ func Test_Config_Vanilla(t *testing.T) {
 
 func Test_Config_Xdg(t *testing.T) {
 	utils.ConfigInitialization.Reset() // Reset sync.Once for the test
-	SetMockKeyring(t)
+	helpers.SetMockKeyring(t)
 	tempDir := t.TempDir()
-	safeSetEnv("HARBOR_CLI_CONFIG", filepath.Join(tempDir, "config.yaml"))
-	safeSetEnv("XDG_CONFIG_HOME", filepath.Join(tempDir, ".config"))
-	safeSetEnv("XDG_DATA_HOME", filepath.Join(tempDir, ".data"))
+	helpers.SafeSetEnv("HARBOR_CLI_CONFIG", filepath.Join(tempDir, "config.yaml"))
+	helpers.SafeSetEnv("XDG_CONFIG_HOME", filepath.Join(tempDir, ".config"))
+	helpers.SafeSetEnv("XDG_DATA_HOME", filepath.Join(tempDir, ".data"))
 	utils.InitConfig("", false)
 	cds := root.RootCmd()
 	err := cds.Execute()
@@ -146,7 +83,7 @@ func Test_Config_Xdg(t *testing.T) {
 
 	currentData, err := utils.GetCurrentHarborData()
 	assert.NoError(t, err, "Expected no error when fetching HarborData")
-	defer ConfigCleanup(t, currentData)
+	defer helpers.ConfigCleanup(t, currentData)
 
 	currentConfig, err := utils.GetCurrentHarborConfig()
 	assert.NoError(t, err, "Expected no error when fetching HarborConfig")
@@ -158,8 +95,8 @@ func Test_Config_Xdg(t *testing.T) {
 
 func Test_Config_Flag(t *testing.T) {
 	tempDir := t.TempDir()
-	data := Initialize(t, tempDir)
-	defer ConfigCleanup(t, data)
+	data := helpers.Initialize(t, tempDir)
+	defer helpers.ConfigCleanup(t, data)
 
 	testConfigFile := filepath.Join(tempDir, "config.yaml")
 	utils.InitConfig(testConfigFile, true)
